backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
psycopg2-binary>=2.9.10
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
Pillow>=10.4.0
